requirements.txt
PyQt5>=5.15.0
opencv-python>=4.5.0
numpy>=1.19.0
requests>=2.25.0
pandas>=1.3.0
pyserial>=3.5
openpyxl>=3.0.0  # 用于pandas导出Excel文件 